refactor(edpaggregator): extract shared VID labeling Cloud Function helpers
VidLabelingDispatcherFunction and VidLabelingMonitorFunction duplicated six
byte-identical private helpers plus a ChannelKey data class and channel cache.
Extract them into a single VidLabelingFunctionHelpers object in the same package:
- channel factory: createPublicChannel, getOrCreateChannel, createInstrumentedChannel
  (+ ChannelKey, channelCache, channelShutdownDuration)
- convertModelLineConfigs
- buildVidLabelerParamsTemplate (carries the required edp_impression_path prefix)
- buildSubpoolAssignerParamsTemplate

Both Functions now delegate to VidLabelingFunctionHelpers; createStorageClient stays
in the dispatcher (it is dispatcher-specific). Behavior-preserving: the moved bodies
are verbatim. Access is confined to edpaggregator subpackages via the package
default_visibility, since Kotlin `internal` cannot cross Bazel module boundaries.

Fixes #4027
